pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "air-pollution-dashboard"
version = "0.1.0"
description = "A dashboard for visualizing air pollution data from multiple cities"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "geopy>=2.4.1",
    "pandas>=2.0.0",
    "pyyaml>=6.0.1",
    "google-cloud-bigquery>=3.11.0",
    "pandas-gbq>=0.19.2",
    "dbt-bigquery>=1.5.0",
    "streamlit>=1.24.0",
    "plotly>=5.15.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
namespaces = false

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 